SubscriberCoordinator refactoring.
